backend/requirements.txt
asgiref==3.7.2
boto3==1.34.68
botocore==1.34.68
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
Django==4.2.11
django-cors-headers==4.3.1
django-filter==24.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-nested-routers==0.93.5
exceptiongroup==1.2.0
idna==3.6
iniconfig==2.0.0
jmespath==1.0.1
numpy==1.26.4
packaging==24.0
pandas==2.2.1
Pandas3==0.0.1
pillow==10.2.0
pluggy==1.4.0
psycopg==3.1.12
psycopg-binary==3.1.12
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytest==8.1.1
pytest-django==4.8.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
requests==2.31.0
s3transfer==0.10.1
scipy==1.12.0
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
